Add: sendChatHandler
