+++ src/modules/patterns.js
@@ -12,5 +12,6 @@ module.exports = {
   },
   chat: {
     chatImgType: RegExp("^(image/png|image/jpg|image/jpeg)$"),
+    chatType: RegExp("^(text|in|out|payment|settlement|account)$"),
   },
 };
