+++ src/services/chats.js
@@ -1,52 +1,16 @@
 const { chatModel, userModel } = require("../modules/stores/mongo");
 const { chatPopulateOption } = require("../modules/populates/chats");
-const validator = require("validator");
 const awsS3 = require("../modules/stores/awsS3");
 
-const { emitChatEvent } = require("./socket.chats");
-
-/**
- * Chat Object의 array가 주어졌을 때 클라이언트에서 처리하기 편한 형태로 Chat Object를 가공합니다.
- * @param {[Object]} chats - Chats Document에 lean과 populate(chatPopulateOption)을 차례로 적용한 Chat Object의 배열입니다.
- * @return {Promise<Array>} {type: String, authorId: String, authorName: String, authorProfileUrl: String, content: string, time: Date}로 이루어진 chat 객체의 배열입니다.
- */
-const transformChatsForRoom = async (chats) => {
-  const chatsToSend = [];
-
-  for (const chat of chats) {
-    // inOutNames 배열(들어오거나 나간 사용자들의 닉네임으로 이루어진 배열)을 생성합니다.
-    chat.inOutNames = [];
-    if (chat.type === "in" || chat.type === "out") {
-      const inOutUserIds = chat.content.split("|");
-      chat.inOutNames = await Promise.all(
-        inOutUserIds.map(async (userId) => {
-          const user = await userModel.findOne({ id: userId }, "nickname");
-          return user.nickname;
-        })
-      );
-    }
-    chatsToSend.push({
-      type: chat.type,
-      authorId: chat.authorId._id,
-      authorName: chat.authorId.nickname,
-      authorProfileUrl: chat.authorId.profileImageUrl,
-      content: chat.content,
-      time: chat.time,
-      isValid: chat.isValid,
-      inOutNames: chat.inOutNames,
-    });
-  }
-  return chatsToSend;
-};
+const { transformChatsForRoom, emitChatEvent } = require("./socket.chats");
 
 const loadRecentChatHandler = async (req, res) => {
   try {
     const io = req.app.get("io");
     const roomId = req.body.roomId;
     const userId = req.session?.userId;
-    const socketId = req.session?.socketId;
     const user = await userModel.findOne({ id: userId });
-    if (!userId || !socketId) {
+    if (!userId) {
       return res.status(403).send("Chat/ : socket did not connected");
     }
     if (!user) {
@@ -62,10 +26,13 @@ const loadRecentChatHandler = async (req, res) => {
 
     if (chats) {
       chats.reverse();
-      io.to(socketId).emit("chat_init", {
+      io.to(`user-${userId}`).emit("chat_init", {
         chats: await transformChatsForRoom(chats),
         roomId,
       });
+      res.json({ result: true });
+    } else {
+      res.json({ result: false });
     }
   } catch (e) {
     res.status(500).send("Chat/ : internal server error");
@@ -77,9 +44,8 @@ const loadBeforeChatHandler = async (req, res) => {
     const io = req.app.get("io");
     const { roomId, lastMsgDate } = req.body;
     const userId = req.session?.userId;
-    const socketId = req.session?.socketId;
     const user = await userModel.findOne({ id: userId });
-    if (!userId || !socketId) {
+    if (!userId) {
       return res
         .status(403)
         .send("Chat/load/before : socket did not connected");
@@ -97,10 +63,13 @@ const loadBeforeChatHandler = async (req, res) => {
 
     if (chats) {
       chats.reverse();
-      io.to(socketId).emit("chat_push_front", {
+      io.to(`user-${userId}`).emit("chat_push_front", {
         chats: await transformChatsForRoom(chats),
         roomId,
       });
+      res.json({ result: true });
+    } else {
+      res.json({ result: false });
     }
   } catch (e) {
     res.status(500).send("Chat/load/before : internal server error");
@@ -112,9 +81,8 @@ const loadAfterChatHandler = async (req, res) => {
     const io = req.app.get("io");
     const { roomId, lastMsgDate } = req.body;
     const userId = req.session?.userId;
-    const socketId = req.session?.socketId;
     const user = await userModel.findOne({ id: userId });
-    if (!userId || !socketId) {
+    if (!userId) {
       return res
         .status(403)
         .send("Chat/load/before : socket did not connected");
@@ -132,10 +100,13 @@ const loadAfterChatHandler = async (req, res) => {
 
     if (chats) {
       chats.reverse();
-      io.to(socketId).emit("chat_push_back", {
+      io.to(`user-${userId}`).emit("chat_push_back", {
         chats: await transformChatsForRoom(chats),
         roomId,
       });
+      res.json({ result: true });
+    } else {
+      res.json({ result: false });
     }
   } catch (e) {
     res.status(500).send("Chat/load/after : internal server error");
@@ -144,8 +115,8 @@ const loadAfterChatHandler = async (req, res) => {
 
 const sendChatHandler = async (req, res) => {
   try {
-    const roomId = req.body.roomId;
-    const chatMessage = req.body.chatMessage;
+    const io = req.app.get("io");
+    const { roomId, type, content } = req.body;
     const userId = req.session?.userId;
     const user = await userModel.findOne({ id: userId });
     if (!userId) {
@@ -154,9 +125,13 @@ const sendChatHandler = async (req, res) => {
     if (!user) {
       return res.status(500).send("Chat/send : internal server error");
     }
-    emitChatEvent(req.app.get("io"), chat.roomId, chat);
+    const result = await emitChatEvent(io, roomId, {
+      type,
+      content,
+      authorId: userId,
+    });
 
-    /* TODO: */
+    res.json({ result });
   } catch (e) {
     res.status(500).send("Chat/send : internal server error");
   }
